feat: 도메인 엔티티 수정
- Base 엔티티 삭제

TODO: audit 공통 관리
